+++ app/helpers/computations_helper.rb
@@ -1,4 +1,8 @@
 # frozen_string_literal: true
 module ComputationsHelper
   include PatientsHelper
+
+  def infrastructure_file_path(path)
+    path&.gsub('download/', 'files/')
+  end
 end