+++ CHANGELOG.md
@@ -22,12 +22,14 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Set plgrid login in OpenId request while regenerating proxy certificate (@mkasztelnik)
 - Long living JWT tokens can be generated using internal API (@mkasztelnik)
 - Resource paths support wildcard characters through UI and API (@dharezlak)
+- User can download patient case computation stdout and stderr (@mkasztelnik)
 
 ### Changed
 - Make jwt pem path configurable though system variable (@mkasztelnik)
 - Externalize JWT token generation from user into separate class (@mkasztelnik)
 - Make PDP URI and access method case insensitive (@mkasztelnik)
 - Improve service URL form placeholder (@mkasztelnik)
+- Externalize PLGrid grant id into configuration file (@mkasztelnik)
 
 ### Deprecated